requirements.txt
appdirs==1.4.4
argcomplete==1.12.0
asgiref==3.2.10
astroid==2.4.2
boto3==1.15.4
botocore==1.18.4
certifi==2020.6.20
cfn-flip==1.2.3
chardet==3.0.4
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.1
Django==3.1
django-cors-headers==3.5.0
django-rest-swagger==2.2.0
django-s3-storage==0.13.4
djangorestframework==3.11.1
drf-yasg==1.17.1
durationpy==0.5
filelock==3.0.12
future==0.18.2
hjson==3.0.2
idna==2.10
importlib-metadata==1.7.0
importlib-resources==3.2.0
inflection==0.5.0
isort==5.4.2
itypes==1.2.0
Jinja2==2.11.2
jmespath==0.10.0
kappa==0.6.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-codec==1.3.2
packaging==20.4
pip-tools==5.3.1
pipenv==2020.8.13
placebo==0.9.0
psycopg2==2.8.5
pylint==2.6.0
PyMySQL==0.10.1
pyparsing==2.4.7
python-dateutil==2.6.1
python-dotenv==0.14.0
python-slugify==4.0.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
simplejson==3.17.2
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.1
tqdm==4.49.0
troposphere==2.6.2
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.10
virtualenv==20.1.0
virtualenv-clone==0.5.4
Werkzeug==0.16.1
wrapt==1.12.1
wsgi-request-logger==0.4.6
zappa==0.52.0
zipp==3.2.0
